I1 tweaks

See merge request comp3350-winter2024/lethalcompany-a01-13!20

+++ app/src/main/java/com/example/easyshopper/logic/ShoppingListHandler.java
@@ -7,11 +7,12 @@ import com.example.easyshopper.objects.Store;
 import com.example.easyshopper.persistence.PricePersistence;
 import com.example.easyshopper.persistence.ShoppingListPersistence;
 
+import java.io.Serializable;
 import java.util.List;
 
 
 //Handles the shopping lists
-public class ShoppingListHandler {
+public class ShoppingListHandler implements Serializable {
     private ShoppingListPersistence shoppingListPersistence = Services.getShoppingListPersistence();
     private PricePersistence pricePersistence = Services.getPricePersistence();
 
@@ -24,23 +25,41 @@ public class ShoppingListHandler {
 
     //add item into the given shopping list
     public void addItemToList(Product newProduct, ShoppingList shoppingList){
+        if(!shoppingListPersistence.shoppingListExists(shoppingList) ||
+                newProduct == null || shoppingList == null) {
+
+            return;
+        }
+
         shoppingList.addProductToCart(newProduct);
         shoppingListPersistence.updateShoppingList(shoppingList);
     }
 
     //create a new shopping list and add to overall shopping list array
     public void createShoppingList(Store store){
+        //validate inputs
+        if(store == null || shoppingListPersistence.listWithStoreExists(store)) {
+            return;
+        }
+
         ShoppingList newList = new ShoppingList(store);
         shoppingListPersistence.addShoppingList(newList);
     }
 
     //remove given shoppingList from overall shopping list array
     public void removeShoppingList(ShoppingList shoppingList){
-        shoppingListPersistence.deleteShoppingList(shoppingList);
+        if(shoppingList != null && shoppingListPersistence.shoppingListExists(shoppingList)) {
+            shoppingListPersistence.deleteShoppingList(shoppingList);
+        }
     }
 
     //remove an item from a shopping list
     public void removeProduct(Product product, ShoppingList shoppingList){
+        if(!shoppingListPersistence.shoppingListExists(shoppingList)||
+                product == null || shoppingList == null) {
+            return;
+        }
+
         shoppingList.removeProductFromCart(product);
         shoppingListPersistence.updateShoppingList(shoppingList);
     }
@@ -48,6 +67,11 @@ public class ShoppingListHandler {
 
     //Gives total price of the ShoppingList
     public double getCartTotal(ShoppingList shoppingList){
+        if(shoppingList == null || !shoppingListPersistence.shoppingListExists(shoppingList))
+        {
+            return -1;
+        }
+
         double total = 0;
         Store store = shoppingList.getStore();
